interfaces/builtin/fuse_support: adjust confinment to allow unmounting fuse filesystems
We let snaps to mount fuse filesystems, but do not allow them to do a umount.
This is a problem because the mount points cannot be cleaned up (fusermount does
not work either because of the confinement), and it leaves behind fuse mounts in
snap namespaces.

Adjust confinement policy to allow umount2 syscall (used by libfuse and
fusermount). Add AppArmor rules that allow umount of fuse filesystems mounted
under snap-specific writable directories.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>