overlord/snapstate: pass context when starting services
This enables aborting of service startup. Useful when some of the services in a
snap are misbehaving and result in systemctl being blocked/waiting for
indefinite amount of this. With the support code in `cmd/snap` the user will be
able to cancel 'starting services' task, resulting in snap being rolled back.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>