cmd/snap: attempt to restore SELinux context of snap user directories
The SELinux policy of snapd allows certain operations on objects that are in
snappy_home_t domain. The trouble is that when a user runs a snap, and ~/snap
does not exist, it gets created using the parent directory context, which is
user_home_t, causing the policy rules fail to match. We could update the policy
to allow manipulation of user_home_t but that would open the whole of user's
home directory. Instead, try to restore the proper context on each run, which
should DTRT for the newly created directory and ones that already exist.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>