overlord/ifacestate: helper API to obtain the state of connections
Add a helper method to query the state of connections. The API returns
connections that are currently established, as well as ones that were explicitly
disconnected. Connection state tells whether the connection was established
automtically, manually or triggered by the gadget.

The primary purpose of the change is to make the information accessible in the
daemon, to satisfy the API calls to /v2/connections endpoint.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>