wrappers: compatibility with pre systemd 231 versions
The range .. syntax was introduced in systemd 231, released on 2016-07-25, while
the last ~ syntax was added in 233, released on 2017-03-01. Thus, any snap with
service timers using a numbered schedule (eg. mon1-tue), or a schedule with
last weekday of a month (eg. mon-fri5), would fail to install since systemd is
unable to interpret the genrated OnCalendar= events in the timer unit.

Fix the schedules by generating explicit event list, with 'x,y,z' syntax that is
supported by the older versions.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>