cmd/snap-confine: inspect CURRENT_TAGS when available
Since version 247 of systemd, the TAGS property is sticky, meaning tags that
were added once stay around. A new property called CURRENT_TAGS was introduced,
which contains the set of tags based on the most recent database update, see:
https://github.com/systemd/systemd/blob/v247/NEWS#L5-L87 When starting a new
process of a given snap, we need to check the CURRENT_TAGS to obtain the set of
tags that show the effect of most recent calls to snap connect/disconnect.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>