cmd/libsnap-confine-private/tool: switch identity only after forking a child
We had a recurring pattern of switching the identity, then calling a
tool, followed by restoring the old identity. Since the tool is called
in a forked child process anyway, there is no clear win to switch the
identity in the parent process, which then requires a followup restore
operation. Move switching of identity to the child process, this
simplifying the code path.

Additional advantage of the change is that PR_SET_KEEPCAPS is no longer
needed, as uid/gid switching is done in the child.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>