cmd/snap-exec, cmd/snap: workaround for Go 1.25 cgroup aware GOMAXPROCS probing
Workaround for cgroup aware setting of GOMAXPROCS introduced in Go.
1.25.

Since snap-exec is a Go binary, it will attempt to probe cgroups, and
read from /proc/self/mountinfo, which is not allowed by the default
AppArmor template for all snaps. This raises a denial like the following:

[  237.760358] audit: type=1400 audit(1757343898.685:370):
   apparmor="DENIED" operation="open" class="file"
   profile="snap.test-snapd-sh.sh"
   name="/proc/20526/mountinfo" pid=20526 comm="snap-exec"
   requested_mask="r" denied_mask="r" fsuid=0 ouid=0

Attempt to workaround this by disabling the probing mechanism and set
GOMAXPROCS to a reasonable value which would otherwise be selected by
the runtime.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>